refactor(gatsby): optimized queries, types and imports
