+++ src/__fixtures__/page.tsx
@@ -1,14 +1,14 @@
 import {
   PaginationInfo,
-  LearnPageContext,
+  LearnTemplateContext,
   NavigationSectionData,
-  BlogPosts,
-  BlogPageData,
-  BlogPageContext,
-  ArticleProps,
+  PostTemplateData,
+  PostTemplateContext,
   NodeReleaseData,
   PageTableOfContents,
-  BlogCategories,
+  BlogCategory,
+  BlogPost,
+  ArticleData,
 } from '../types';
 import mockMDXBodyContent from './mockMDXBodyContent';
 
@@ -59,7 +59,7 @@ export const createNavigationSectionData = (): NavigationSectionData =>
     },
   } as NavigationSectionData);
 
-export const createLearnPageData = (): ArticleProps => ({
+export const createLearnPageData = () => ({
   data: {
     articleCurrentLanguage: {
       body: mockMDXBodyContent,
@@ -71,7 +71,7 @@ export const createLearnPageData = (): ArticleProps => ({
       fields: {
         authors: ['test-user1', 'test-user2'],
       },
-    },
+    } as ArticleData,
   },
 });
 
@@ -112,26 +112,51 @@ export const createNodeReleasesData = (): NodeReleaseData[] =>
     },
   ] as NodeReleaseData[];
 
-export const createLearnPageContext = (): LearnPageContext =>
+export const createLearnPageContext = (): LearnTemplateContext =>
   ({
     slug: 'test-slug',
     relativePath: 'test-path',
     next: createPaginationInfo(),
     previous: createPaginationInfo(),
     navigationData: createNavigationSectionData(),
-  } as LearnPageContext);
+  } as LearnTemplateContext);
 
-export const createBlogPageContext = (): BlogPageContext => ({
+export const createBlogPageContext = (): PostTemplateContext => ({
   slug: 'test-slug',
   relativePath: 'test-path',
   next: createPaginationInfo(),
   previous: createPaginationInfo(),
   navigationData: createNavigationSectionData(),
+  recent: [
+    {
+      node: {
+        frontmatter: {
+          title: 'title-mock',
+          category: {
+            name: 'category-mock',
+            slug: 'category-mock-slug',
+          },
+          blogAuthors: [
+            {
+              id: 'id-mock',
+              name: 'name-mock',
+              website: 'url-mock',
+            },
+          ],
+        },
+        fields: {
+          date: '11/11/2022',
+          slug: 'slug-mock',
+          readingTime: { text: 'text-mock' },
+        },
+      },
+    },
+  ],
 });
 
 export const createBlogData = (): {
-  posts: BlogPosts;
-  categories: BlogCategories;
+  posts: { edges: BlogPost[] };
+  categories: { edges: BlogCategory[] };
 } => ({
   posts: {
     edges: [
@@ -166,8 +191,8 @@ export const createBlogData = (): {
   },
 });
 
-export const createBlogPageData = (): BlogPageData => ({
-  blog: {
+export const createBlogPageData = (): PostTemplateData => ({
+  mdx: {
     body: mockMDXBodyContent,
     excerpt: 'excerpt-mock',
     frontmatter: {
@@ -182,36 +207,9 @@ export const createBlogPageData = (): BlogPageData => ({
     },
     fields: { slug: 'slug-mock', date: '11/11/2022' },
   },
-  recent: {
-    edges: [
-      {
-        node: {
-          frontmatter: {
-            title: 'title-mock',
-            category: {
-              name: 'category-mock',
-              slug: 'category-mock-slug',
-            },
-            blogAuthors: [
-              {
-                id: 'id-mock',
-                name: 'name-mock',
-                website: 'url-mock',
-              },
-            ],
-          },
-          fields: {
-            date: '11/11/2022',
-            slug: 'slug-mock',
-            readingTime: { text: 'text-mock' },
-          },
-        },
-      },
-    ],
-  },
 });
 
-export const createGeneralPageData = (): ArticleProps => ({
+export const createGeneralPageData = () => ({
   data: {
     articleCurrentLanguage: {
       fields: {
@@ -223,11 +221,11 @@ export const createGeneralPageData = (): ArticleProps => ({
       },
       body: mockMDXBodyContent,
       tableOfContents: mockTableOfContents,
-    },
+    } as ArticleData,
   },
 });
 
-export const createResourcesData = (): ArticleProps => ({
+export const createResourcesData = () => ({
   data: {
     articleCurrentLanguage: {
       fields: {
@@ -239,11 +237,11 @@ export const createResourcesData = (): ArticleProps => ({
       },
       body: mockMDXBodyContent,
       tableOfContents: mockTableOfContents,
-    },
+    } as ArticleData,
   },
 });
 
-export const createPrivacyData = (): ArticleProps => ({
+export const createPrivacyData = () => ({
   data: {
     articleCurrentLanguage: {
       fields: {
@@ -255,6 +253,6 @@ export const createPrivacyData = (): ArticleProps => ({
       },
       body: mockMDXBodyContent,
       tableOfContents: mockTableOfContents,
-    },
+    } as ArticleData,
   },
 });
