+++ src/components/ApiComponents/Components/index.tsx
@@ -2,5 +2,6 @@ export { default as ApiLink } from './ApiLink';
 export { default as Stability } from './Stability';
 export { default as SourceLink } from './SourceLink';
 export { default as Changes } from './Changes';
+export { default as VersionSelector } from './VersionSelector';
 export { default as Span } from './Span';
 export * from './ApiHeading';
