+++ src/components/RecentPosts/__tests__/index.test.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import RecentPosts from '..';
-import { createBlogPageData } from '../../../__fixtures__/page';
+import { createBlogPageContext } from '../../../__fixtures__/page';
 
 describe('RecentPosts component', (): void => {
   it('should render correctly', (): void => {
-    const {
-      recent: { edges: recentPosts },
-    } = createBlogPageData();
+    const { recent } = createBlogPageContext();
 
-    const { container } = render(<RecentPosts posts={recentPosts} />);
+    const { container } = render(<RecentPosts posts={recent} />);
 
     expect(container).toMatchSnapshot();
   });
