+++ src/styles/templates/api.module.scss
@@ -1,4 +1,4 @@
-@use '../styles/containers.scss';
+@use '../containers.scss';
 
 .api-container {
   @extend .grid-container;
