+++ src/styles/templates/post.module.scss
@@ -1,4 +1,4 @@
-.blog-container {
+.post-container {
   grid-gap: var(--space-24);
 
   article {
