+++ src/types/markdown.ts
@@ -1,0 +1,14 @@
+export interface TableOfContentsItem {
+  title: string;
+  url: string;
+  items?: TableOfContentsItem[];
+}
+
+export interface PageTableOfContents {
+  items: TableOfContentsItem[];
+}
+
+export interface PaginationInfo {
+  slug: string;
+  title: string;
+}
