+++ src/types/pages/blog.ts
@@ -1,0 +1,7 @@
+import { BlogCategory, BlogPost } from '../blog';
+
+export interface BlogTemplateContext {
+  posts: BlogPost[];
+  categories: BlogCategory[];
+  category: BlogCategory['node'];
+}
