+++ src/types/pages/index.ts
@@ -1,0 +1,9 @@
+export interface GenericPageContext {
+  intlMessages: Record<string, string>;
+  locale: string;
+}
+
+export * from './api';
+export * from './learn';
+export * from './post';
+export * from './blog';
