+++ util-node/createBlogQuery.js
@@ -1,0 +1,43 @@
+module.exports = `
+{
+  pages: allMdx(
+    filter: { fileAbsolutePath: { regex: "/blog/" } }
+    sort: { fields: fields___date, order: DESC }
+  ) {
+    edges {
+      node {
+        id
+        fileAbsolutePath
+        parent {
+          ... on File {
+            relativePath
+          }
+        }
+        frontmatter {
+          title
+          blogAuthors {
+            id
+            name
+          }
+        }
+        fields {
+          slug
+          categoryName
+          date(formatString: "MMMM DD, YYYY")
+          readingTime {
+            text
+          }
+        }
+      }
+    }
+  }
+  categories: allCategoriesYaml {
+    edges {
+      node {
+        slug
+        name
+      }
+    }
+  }
+}
+`;
